move feed store and local feed loader into its own files
