+++ FeedFeature/Feed Cache/FeedStore.swift
@@ -1,0 +1,16 @@
+//
+//  FeedStore.swift
+//  FeedFeature
+//
+//  Created by Srinivasan Rajendran on 2020-12-22.
+//
+
+import Foundation
+
+public protocol FeedStore {
+    typealias DeletionCompletion = (Error?) -> Void
+    typealias InsertionCompletion = (Error?) -> Void
+
+    func deleteCachedFeed(completion: @escaping DeletionCompletion)
+    func insertFeed(items: [FeedItem], timestamp: Date, completion: @escaping InsertionCompletion)
+}
